requirements.txt
APScheduler==3.6.3
schedule==0.6.0
requests==2.24.0
botoy==0.0.15
pytz==2020.1
beautifulsoup4==4.9.3
Pillow==8.1.0
